ProductEntity tests - added Delete. Changes in ProductRepository - Long instead of Integer (as ProductEntity Id)

+++ src/main/java/com/kodilla/ecommercee/repository/ProductRepository.java
@@ -8,5 +8,7 @@ import javax.transaction.Transactional;
 
 @Transactional
 @Repository
-public interface ProductRepository extends CrudRepository<ProductEntity, Integer> {
+public interface ProductRepository extends CrudRepository<ProductEntity, Long> {
+
 }
+
